requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
asyncer==0.0.1
attrs==24.3.0
beanie==1.28.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
clarabel==0.9.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cssselect==1.2.0
cvxpy==1.6.0
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
dnspython==2.7.0
ecos==2.0.14
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.1.0
fake-useragent==2.0.3
fastapi==0.115.6
fastapi-cli==0.0.6
feedparser==6.0.11
fonttools==4.55.3
frozendict==2.4.6
frozenlist==1.5.0
h11==0.12.0
html5lib==1.1
httpcore==0.14.7
httptools==0.6.4
httpx==0.21.3
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy-model==0.2.0
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
motor==3.6.0
multidict==6.1.0
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.2.0
osqp==0.6.7.post3
packaging==24.2
pandas==2.2.3
pandas-datareader==0.10.0
parse==1.20.2
parso==0.8.4
peewee==3.17.8
pendulum==2.1.2
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==1.10.19
pydantic_core==2.27.1
pyee==11.1.1
Pygments==2.18.0
pymongo==4.9.2
pyparsing==3.2.0
pyportfolioopt==1.5.6
pyppeteer==2.0.0
pyquery==2.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
pytzdata==2020.1
PyYAML==6.0.2
pyzmq==26.2.0
qdldl==0.1.7.post5
requests==2.32.3
requests-html==0.10.0
rfc3986==1.5.0
rich==13.9.4
rich-toolkit==0.12.0
scipy==1.14.1
scs==3.2.7.post2
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
tenacity==8.5.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.32.1
w3lib==2.2.1
watchfiles==1.0.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4
yahoo-fin==0.8.9.1
yarl==1.18.3
zipp==3.21.0
